+++ server/server.go
@@ -35,6 +35,7 @@ import (
 	"github.com/influxdata/kapacitor/services/dns"
 	"github.com/influxdata/kapacitor/services/ec2"
 	"github.com/influxdata/kapacitor/services/file_discovery"
+	"github.com/influxdata/kapacitor/services/fluxtask"
 	"github.com/influxdata/kapacitor/services/gce"
 	"github.com/influxdata/kapacitor/services/hipchat"
 	"github.com/influxdata/kapacitor/services/httpd"
@@ -76,9 +77,11 @@ import (
 	"github.com/influxdata/kapacitor/services/udp"
 	"github.com/influxdata/kapacitor/services/victorops"
 	"github.com/influxdata/kapacitor/services/zenoss"
+	"github.com/influxdata/kapacitor/task/taskmodel"
 	"github.com/influxdata/kapacitor/uuid"
 	"github.com/influxdata/kapacitor/waiter"
 	"github.com/pkg/errors"
+	"go.uber.org/zap/zapcore"
 )
 
 const clusterIDFilename = "cluster.id"
@@ -117,6 +120,8 @@ type Server struct {
 
 	TaskMaster       *kapacitor.TaskMaster
 	TaskMasterLookup *kapacitor.TaskMasterLookup
+
+	FluxTaskService taskmodel.TaskService
 
 	LoadService           *load.Service
 	SideloadService       *sideload.Service
@@ -246,6 +251,10 @@ func New(c *Config, buildInfo BuildInfo, diagService *diagnostic.Service) (*Serv
 
 	if err := s.appendInfluxDBService(); err != nil {
 		return nil, errors.Wrap(err, "influxdb service")
+	}
+
+	if err := s.appendFluxTaskService(); err != nil {
+		return nil, errors.Wrap(err, "fluxtask service")
 	}
 
 	if err := s.appendLoadService(); err != nil {
@@ -399,6 +408,17 @@ func (s *Server) initAlertService() {
 
 func (s *Server) appendAlertService() {
 	s.AppendService("alert", s.AlertService)
+}
+
+func (s *Server) appendFluxTaskService() error {
+	// TODO: hook into correct logging level instead of assuming info
+	logger := s.DiagService.NewZapLogger(zapcore.InfoLevel)
+	srv := fluxtask.New(s.config.FluxTask, logger)
+	srv.HTTPDService = s.HTTPDService
+	srv.InfluxDBService = s.InfluxDBService
+	srv.StorageService = s.StorageService
+	s.AppendService("fluxtask", srv)
+	return nil
 }
 
 func (s *Server) appendTesterService() {